requirements.txt
django==3.2.2
djangorestframework == 3.14.0
Authlib
